pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "suite-automacao-qa"
version = "1.0.0"
description = "Suite de automação de testes: API Petstore + Web SauceDemo"
requires-python = ">=3.11"

dependencies = [
    "requests>=2.31.0",
    "selenium>=4.18.0",
    "jsonschema>=4.21.1",
    "faker>=24.0.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "pytest>=8.1.0",
    "pytest-html>=4.1.1",
    "allure-pytest>=2.13.5",
    "pytest-xdist>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "pre-commit>=3.7.0",
    "pip-tools>=7.4.0",
]

[tool.setuptools]
packages = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Teste*"]
python_functions = ["teste_*"]
markers = [
    "api: marks tests as API layer",
    "web: marks tests as Web/UI layer",
    "smoke: marks tests as smoke suite",
    "regression: marks tests as regression suite",
]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
